Using var when possible
