+++ src/SheetToObjects.Lib/Configuration/MappingConfigBuilder.cs
@@ -2,9 +2,6 @@
 using System.Linq;
 using System.Reflection;
 using SheetToObjects.Lib.Attributes;
-using SheetToObjects.Lib.Attributes.MappingType;
-using SheetToObjects.Lib.Attributes.Rules;
-using SheetToObjects.Lib.Configuration.ColumnMappings;
 
 namespace SheetToObjects.Lib.Configuration
 {
@@ -37,9 +34,9 @@ namespace SheetToObjects.Lib.Configuration
 
         private void InitByAtributes()
         {
-            Type objType = typeof(TModel);
+            var objType = typeof(TModel);
 
-            SheetToObjectConfig sheetToConfigAttribute = objType.GetCustomAttributes().OfType<SheetToObjectConfig>().FirstOrDefault();
+            var sheetToConfigAttribute = objType.GetCustomAttributes().OfType<SheetToObjectConfig>().FirstOrDefault();
             if (sheetToConfigAttribute != null)
             {
                 _mappingConfig.HasHeaders = sheetToConfigAttribute.SheetHasHeaders;
@@ -47,62 +44,6 @@ namespace SheetToObjects.Lib.Configuration
             }
 
             new ColumnsMappingBuilder<TModel>(_mappingConfig);
-        }
-    }
-
-    public class ColumnsMappingBuilder<TModel>
-    {
-        private readonly MappingConfig _mappingConfig;
-
-        public ColumnsMappingBuilder(MappingConfig mappingConfig)
-        {
-            _mappingConfig = mappingConfig;
-            InitByAttributes();
-        }
-
-        public ColumnsMappingBuilder<TModel> Add(Func<ColumnMappingBuilder<TModel>, ColumnMapping> columnMappingBuilderFunc)
-        {
-            var columnMapping = columnMappingBuilderFunc(new ColumnMappingBuilder<TModel>());
-
-            _mappingConfig.ColumnMappings.RemoveAll(c => c.PropertyName == columnMapping.PropertyName);
-            _mappingConfig.ColumnMappings.Add(columnMapping);
-
-            return this;
-        }
-
-        private void InitByAttributes()
-        {
-            Type objType = typeof(TModel);
-
-            foreach (var property in objType.GetProperties())
-            {
-                bool columnIsMappedByAttribute = false;
-                var mappingConfigBuilder = new ColumnMappingBuilder<TModel>();
-                var attributes = property.GetCustomAttributes().ToList();
-
-                if(attributes.OfType<IgnorePropertyMapping>().Any())
-                    continue;
-                
-                foreach (var mappingAttribute in attributes.OfType<IMappingAttribute>())
-                {
-                    mappingAttribute.SetColumnMapping<TModel>(mappingConfigBuilder);
-                    columnIsMappedByAttribute = true;
-                }
-                
-                foreach (var attribute in attributes)
-                {
-                    if (attribute is IRuleAttribute ruleAttribute)
-                    {
-                        mappingConfigBuilder.AddRule(ruleAttribute.GetRule());
-                    }
-                }
-
-                if (columnIsMappedByAttribute || _mappingConfig.AutoMapProperties)
-                {
-                    _mappingConfig.ColumnMappings.Add(mappingConfigBuilder.MapTo(property));
-                }
-            }
-
         }
     }
 }
