+++ src/SheetToObjects.Specs/Lib/Validation/RequiredRuleSpecs.cs
@@ -20,7 +20,7 @@ namespace SheetToObjects.Specs.Lib.Validation
         [Fact]
         public void GivenValidatingRequiredValue_WhenValueIsEmpty_ValidationFails()
         {
-            string value = string.Empty;
+            var value = string.Empty;
 
             var result = new RequiredRule().Validate(value);
 
@@ -31,7 +31,7 @@ namespace SheetToObjects.Specs.Lib.Validation
         [Fact]
         public void GivenValidatingRequiredValueWithAllowedWhitespace_WhenValueIsWhiteSpace_ValidationFails()
         {
-            string value = " ";
+            var value = " ";
 
             var result = new RequiredRule().AllowWhiteSpace().Validate(value);
 
@@ -41,7 +41,7 @@ namespace SheetToObjects.Specs.Lib.Validation
         [Fact]
         public void GivenValidatingRequiredValueWithoutAllowedWhitespace_WhenValueIsWhiteSpace_ValidationFails()
         {
-            string value = " ";
+            var value = " ";
 
             var result = new RequiredRule().Validate(value);
 
@@ -52,7 +52,7 @@ namespace SheetToObjects.Specs.Lib.Validation
         [Fact]
         public void GivenValidatingRequiredValue_WhenValueIsSet_ValidationIsSuccessful()
         {
-            int value = 42;
+            var value = 42;
 
             var result = new RequiredRule().Validate(value.ToString());
